refactor(planner): introduce IRecipePlanner port (#88, phase A)
Structural prep for #88. No behavior change — every existing test
passes byte-for-byte.

- Introduce IRecipePlanner port in ERP.Application (top-level,
  mirrors IFactoryStateProvider placement).
- Move existing recursive expansion logic from the static
  PlanProductionHandler into RecursiveRecipePlanner : IRecipePlanner
  (instance class, takes ICatalogProvider in ctor).
- PlanProductionHandler shrinks to a Wolverine entry-point shim
  that delegates to the injected planner.
- Register IRecipePlanner → RecursiveRecipePlanner in
  AddErpInfrastructure (singleton).
- Rename PlanProductionHandlerTests → RecursiveRecipePlannerTests;
  tests instantiate the planner directly rather than calling the
  static shim. Git tracks this as a rename.

Phase B (OR-Tools LP adapter behind a Planner:Engine config flag)
follows in a separate PR. OR-Tools native deps already verified to
load on both macOS + self-hosted Linux runner via PoC.
